Updated circleci debug and release jobs to explicitly build Play flavors
Changed the gradlew build targets assembleRelease to assemblePlayRelease
and assembleDebug to assemblePlayDebug, because the old targets causes
files for the free builds to get compiled when not needed. It's
unnecessary and also done without -PfreeBuild which gives build errors.

These are also the targets used in makeRelease.sh, so the workflow
should better match the expected.